Fix an infinite readline() loop.
At least I am in the habit of removing the boilerplate text, just so
that I'm absolutely sure that the commit message won't contain
anything other than what I've written (e.g. if I accidentally deleted
a character or two from the boilerplate text).  Thus, the end of
the commit message may be reached before the separator is found.