Prevent gitless from following symlinks
Add a check for symlinks when we check if a path is a directory and treat the symlink as a normal
file. Gitless will NOT follow the symlinks, because it could pull in the .git/ directory, files
outside of the repository or, in the worst case, the entire filesystem. This follows the behavior of
git. Git will commit symlinks into the repository and recreate the links on filesystems that support
it when the file is downloaded.